chore(github): migrate travis ci/cd to gh actions - I4 (#352)
