--- .travis.yml
@@ -1,33 +1,0 @@
-language: node_js
-node_js:
-  - '10'
-matrix:
-  include:
-    - os: linux
-      dist: trusty
-    - os: windows
-before_install: |
-  set -e
-  export PATH="./$HOME/node_modules/.bin:$PATH"
-  npm install -g @alrra/travis-scripts
-  date
-install: |
-  set -e
-  date
-  npm install -g lerna@^3.20.2
-  npm install 2>&1
-  date
-script: npm run test 2>&1
-deploy:
-  provider: script
-  script: 'if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then bash ./.travis/deploy.sh; fi'
-  skip_cleanup: true
-  on:
-    all_branches: true
-    repo: accordproject/markdown-transform
-    condition: "$TRAVIS_OS_NAME = linux"
-after_failure: tail -n +1 -- /home/travis/.npm/_logs/*-debug.log
-after_success: if [ $TRAVIS_OS_NAME = 'linux' ] && [ $TRAVIS_REPO_SLUG = 'accordproject/markdown-transform' ]; then npm run coverage; fi
-cache:
-  directories:
-  - node_modules
