+++ lerna.json
@@ -4,10 +4,5 @@
     "packages/*"
   ],
   "version": "0.12.12",
-  "hoist": true,
-  "command": {
-    "publish": {
-      "allowBranch": "master"
-    }
-  }
+  "hoist": true
 }
