+++ packages/markdown-cicero/.eslintignore
@@ -3,5 +3,4 @@ node_modules
 out
 umd
 test/data
-.travis
 scripts
