+++ packages/markdown-cicero/package.json
@@ -61,7 +61,6 @@
     "src": [
       "**/*.js",
       "!./scripts/**/*",
-      "!./.travis/**/*",
       "!./test/**/*",
       "!./coverage/**/*",
       "!./node_modules/**/*",
