+++ packages/markdown-html/.eslintignore
@@ -5,5 +5,4 @@ umd
 lib
 test/data
 src/grammars
-.travis
 scripts
