+++ packages/markdown-html/package.json
@@ -89,7 +89,6 @@
     "src": [
       "**/*.js",
       "!./scripts/**/*",
-      "!./.travis/**/*",
       "!./test/**/*",
       "!./coverage/**/*",
       "!./node_modules/**/*",
