+++ packages/markdown-pdf/.eslintignore
@@ -5,6 +5,5 @@ umd
 lib
 test/data
 src/grammars
-.travis
 scripts
 src/domstubs.js
