+++ packages/markdown-slate/package.json
@@ -57,7 +57,6 @@
     "src": [
       "**/*.js",
       "!./scripts/**/*",
-      "!./.travis/**/*",
       "!./test/**/*",
       "!./coverage/**/*",
       "!./node_modules/**/*",
