+++ packages/markdown-transform/package.json
@@ -69,7 +69,6 @@
     "src": [
       "**/*.js",
       "!./scripts/**/*",
-      "!./.travis/**/*",
       "!./test/**/*",
       "!./coverage/**/*",
       "!./node_modules/**/*",
